respond to latest api changes
