+++ src/dotnet-interactive-vscode/insiders/src/notebookSerializers.ts
@@ -86,7 +86,7 @@ export class DotNetDibNotebookSerializer extends DotNetNotebookSerializer {
 
     static registerNotebookSerializer(context: vscode.ExtensionContext, notebookType: string, clientMapper: ClientMapper, outputChannel: OutputChannelAdapter) {
         const serializer = new DotNetDibNotebookSerializer(clientMapper, outputChannel);
-        const notebookSerializer = vscode.notebook.registerNotebookSerializer(notebookType, serializer);
+        const notebookSerializer = vscode.notebooks.registerNotebookSerializer(notebookType, serializer);
         context.subscriptions.push(notebookSerializer);
     }
 }
@@ -98,7 +98,7 @@ export class DotNetLegacyNotebookSerializer extends DotNetNotebookSerializer {
 
     static registerNotebookSerializer(context: vscode.ExtensionContext, notebookType: string, clientMapper: ClientMapper, outputChannel: OutputChannelAdapter) {
         const serializer = new DotNetLegacyNotebookSerializer(clientMapper, outputChannel);
-        const notebookSerializer = vscode.notebook.registerNotebookSerializer(notebookType, serializer);
+        const notebookSerializer = vscode.notebooks.registerNotebookSerializer(notebookType, serializer);
         context.subscriptions.push(notebookSerializer);
     }
 }
