+++ src/dotnet-interactive-vscode/stable/package.json
